实验三

+++ huadb/src/executors/delete_executor.cpp
@@ -1,5 +1,7 @@
 #include "executors/delete_executor.h"
 
+#include "common/exceptions.h"
+
 namespace huadb {
 
 DeleteExecutor::DeleteExecutor(ExecutorContext &context, std::shared_ptr<const DeleteOperator> plan,
@@ -15,9 +17,15 @@ std::shared_ptr<Record> DeleteExecutor::Next() {
     return nullptr;
   }
   uint32_t count = 0;
+  if (!context_.GetLockManager().LockTable(context_.GetXid(), LockType::IX, table_->GetOid())) {
+    throw DbException("Cannot acquire table lock");
+  }
   while (auto record = children_[0]->Next()) {
     // 通过 context_ 获取正确的锁，加锁失败时抛出异常
     // LAB 3 BEGIN
+    if (!context_.GetLockManager().LockRow(context_.GetXid(), LockType::X, table_->GetOid(), record->GetRid())) {
+      throw DbException("Cannot acquire row lock");
+    }
     table_->DeleteRecord(record->GetRid(), context_.GetXid(), true);
     count++;
   }
